Make lexer recognize docstrings tokens for gates.
Documentation strings in OPENQASM (docstrings) are challenging since
their syntax is the same as the comment syntax. In the specification,
it says the first comment of a gate can contain documentation. However,
when referred to the example, the documentation is in the comment just
above the gate line.

Following, qasmsim assumes the docstring of a gate is the literal
content of the line comments right before a gate declaration. For
instance:

```
OPENQASM 2.0;

// This is not a docstring.

// This multiline comment
// is not a docstring either.

// This is a single-line docstring.
gate id q {}

// This multiline comment
// is also the docstring of the following gate.
gate identity q {}
```

The modifications in the lexer of this PR attempt to produce
the following sequence of tokens:

```
DocStr { repr: " This is a..." }
Gate
id { repr: "id" }
...
DocStr { repr: " This multiline comment\n..." }
Gate
id { repr: "identity" }
...
```

So, only the comments right before a `Gate` token are indeed `DocStr`
tokens. These tokens can be seen as a sort of string literal. Notice
how the blank spaces right after the comment mark `//` are preserved
as well as the new lines at the end of the comments.

The dependency between the `DocStr` and `Gate` tokens (the former
can only appear right before the latter) is a form os context-sensitive
lexing
(https://en.wikipedia.org/wiki/Lexical_analysis#Context-sensitive_lexing)
similar to what happens in Python when concatenating several strings in
multiple lines.

For OPENQASM docstrings, it requires a buffer holding a potential on-going
docstring. All comments are potential docstrings so, when a comment is
found, the lexer enters in a `Mode::Comment` mode that adds the content
of the comment, including the new-line character if it exists, to a
docstring buffer.

If the token right after a comment is `Gate` token, that iteration of
the lexer emits the on-going docstring inside a `DocStr` token (and
discard it) and schedule the emission of the `Gate` token for right
after.

The on-going docstring is discarded when detecting a different token
or a new-line.